robot/requirements.txt
RPi.GPIO==0.7.1
simple-pid==1.0.1
adafruit-circuitpython-tcs34725==3.3.10
websocket-client==1.3.2
requests==2.27.1